Finished restructure
