+++ parking-lot/src/pages/Lot.js
@@ -5,23 +5,6 @@ import useRedirect from '../hooks/useRedirect';
 import LotHeader from '../components/headers/LotHeader';
 import EntryList from '../components/lists/EntryList';
 
-const exampleLot = {
-  name: "Hello World",
-  description: "Test Lot to see if layout is working",
-  author: "Developer",
-  members: 3,
-  code: "AVC123ab",
-  entries: [{
-    name: "Example 1",
-    author: "Student",
-    comments: 3
-  }, {
-    name: "Example 2",
-    author: "Student",
-    comments: 0
-  }]
-}
-
 export default function Lot({ isLoggedIn }) {
 
   const shouldRedirect = isLoggedIn ? false : true;
